refactor files workspace upload flow and domain utilities
