+++ package.json
@@ -10,7 +10,8 @@
     "supabase:types": "npx supabase gen types --project-id $SUPABASE_PROJECT_REF --schema public --lang typescript > lib/supabase/database.types.ts",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
-    "test:e2e:headed": "playwright test --headed"
+    "test:e2e:headed": "playwright test --headed",
+    "test:unit": "vitest run components/files/__tests__"
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.57.4",
@@ -26,6 +27,7 @@
     "@types/react-dom": "^19.0.4",
     "eslint": "^9.24.0",
     "eslint-config-next": "^15.3.3",
-    "typescript": "^5.8.3"
+    "typescript": "^5.8.3",
+    "vitest": "^3.2.4"
   }
 }
